[streamer] Fix musl build
msghdr cannot be directly instantiated on musl due to private padding
fields (see rust-lang/libc#2344). The recommended solution is to use
`mem::zeroed` to initialize the struct then set the fields as needed.

While we could use the same solution for both glibc and musl, we use a
conditional compilation directive to avoid using unsafe code when it's
not required.